\name{splits-defunct}
\alias{splits-defunct}
\alias{compare}

\title{Defunct functions in packaga splits}

\description{
Following function(s) has been removed from \bold{splits} and no longer available.
}

\usage{
compare(...)
}

\details{
\code{compare} has been removed bacause the chi-square test between the single and multiple threshold models is not valid. (Their numbers of parameters are identical. See reference).
}

\references{
Fujisawa, T and Barraclough T.G. in press. Delimiting species using single locus data and the generalized mixed Yule coalescent approach. Systematic Biology.
}
